refactor: 타입 코드 분리
