+++ apps/web/src/components/common/auction-item-card.tsx
@@ -8,7 +8,7 @@ import { Badge } from '@/components/ui/badge';
 
 import useCountdown from '@/hooks/common/useCountdown';
 
-import { AuctionItemProps } from '@/lib/types/auction';
+import { AuctionItemProps } from '@/types/auction';
 
 const renderTimeBadge = ({ hours, minutes }: { hours: string; minutes: string }) => {
   const timeBadgeStyle = hours === '00' && minutes === '00' ? 'closed' : 'tertiary';
