+++ apps/web/src/lib/api/auction.ts
@@ -1,4 +1,4 @@
-import { AuctionListResponse } from '@/types/auction';
+import { AuctionListResponse } from '@/types/auction-prisma';
 import apiClient from './client';
 
 // 경매 목록 조회
