+++ apps/web/src/lib/types/auction-prisma.ts
@@ -1,15 +1,3 @@
-interface AuctionItemProps {
-  id: string;
-  title: string;
-  status: '경매중' | '경매종료';
-  originalPrice: number;
-  currentPrice: number;
-  deadline: string;
-  thumbnailUrl: string;
-}
-
-export type { AuctionItemProps };
-
 import type { Prisma } from '@repo/db';
 
 // 경매 목록 아이템
